requirements.txt
jinja2
pyyaml
telethon
tgcrypto
